# openapi3.yaml
openapi: 3.0.1
info:
  title: CRUD API
  description: basic template for map colonies service
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

paths:
  /anotherResource:
    get:
      operationId: getAnotherResource
      tags:
        - anotherResource
      summary: gets the resource
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/anotherResource'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /products:
    get:
      operationId: getProducts
      tags:
        - products
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum:
              - raster
              - rasterized_vector
              - tiles3d
              - QMesh
      summary: gets the resource
      responses:
        200:
          description: A JSON array of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Products'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      operationId: createProduct
      tags:
        - products
      summary: creates a new record of type product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        201:
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /products/{id}:
    put:
      operationId: updateProduct
      tags:
        - products
      summary: update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProduct'
      responses:
        200:
          description: updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      operationId: deleteProduct
      tags:
        - products
      summary: delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedProductResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
security:
  - {}
components:
  schemas:
    GeoJsonPolygon:
      type: object
      required:
        - type
        - coordinates
      properties:
        type:
          type: string
          enum:
            - Polygon
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              minItems: 2
              maxItems: 2
              items:
                type: number
    error:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 48
        description:
          type: string
          nullable: true
        bounding_polygon:
          $ref: '#/components/schemas/GeoJsonPolygon'
        consumption_link:
          type: string
          nullable: true
        resolution_best:
          type: number
          nullable: true
          format: double
        min_zoom:
          type: integer
          nullable: true
          minimum: 0
        max_zoom:
          type: integer
          nullable: true
          minimum: 0
        type:
          type: string
          enum:
            - raster
            - rasterized_vector
            - tiles3d
            - QMesh
        consumption_protocol:
          type: string
          enum:
            - WMS
            - WMTS
            - XYZ
            - 3D Tiles
    Products:
      type: array
      items:
        $ref: '#/components/schemas/Product'
    UpdateProduct:
      type: object
      properties:
        name:
          type: string
          maxLength: 48
        description:
          type: string
          nullable: true
        bounding_polygon:
          $ref: '#/components/schemas/GeoJsonPolygon'
        consumption_link:
          type: string
          nullable: true
        type:
          type: string
          enum: [raster, rasterized_vector, tiles3d, QMesh]
        consumption_protocol:
          type: string
          enum: [WMS, WMTS, XYZ, '3D Tiles']
        resolution_best:
          type: number
        min_zoom:
          type: integer
        max_zoom:
          type: integer
    DeletedProductResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Product'
    anotherResource:
      type: object
      required:
        - kind
        - isAlive
      properties:
        kind:
          type: string
        isAlive:
          type: boolean